fix + simplify more tests
