+++ packages/synapse-react-client/test/utils/hooks/SynapseAPI/useEntityBundle.test.tsx
@@ -1,21 +1,9 @@
-import { renderHook } from '@testing-library/react'
-import React from 'react'
-import { QueryClient } from 'react-query'
+import { renderHook, waitFor } from '@testing-library/react'
 import useGetEntityBundle from '../../../../src/synapse-queries/entity/useEntityBundle'
 import { EntityBundle } from '@sage-bionetworks/synapse-types'
 import { MOCK_CONTEXT_VALUE } from '../../../../mocks/MockSynapseContext'
-import FullContextProvider from '../../../../src/utils/context/FullContextProvider'
-
-const queryClient = new QueryClient()
-
-const wrapper = (props: { children: React.ReactChildren }) => (
-  <FullContextProvider
-    synapseContext={MOCK_CONTEXT_VALUE}
-    queryClient={queryClient}
-  >
-    {props.children}
-  </FullContextProvider>
-)
+import { createWrapper } from '../../../testutils/TestingLibraryUtils'
+import SynapseClient from '../../../../src/synapse-client'
 
 const expected: EntityBundle = {
   entity: {
@@ -25,22 +13,17 @@ const expected: EntityBundle = {
   },
 }
 
-const SynapseClient = require('../../../../src/synapse-client/SynapseClient')
-SynapseClient.getEntityBundleV2 = jest.fn().mockResolvedValue(expected)
+jest.spyOn(SynapseClient, 'getEntityBundleV2').mockResolvedValue(expected)
 
 describe('useEntityBundle functionality', () => {
-  beforeEach(() => {
-    queryClient.clear()
-  })
-
   it('correctly calls SynapseClient', async () => {
     const entityId = 'syn123'
 
-    const { result, waitFor } = renderHook(() => useGetEntityBundle(entityId), {
-      wrapper,
+    const { result } = renderHook(() => useGetEntityBundle(entityId), {
+      wrapper: createWrapper(),
     })
 
-    await waitFor(() => result.current.isSuccess)
+    await waitFor(() => expect(result.current.isSuccess).toBe(true))
     expect(SynapseClient.getEntityBundleV2).toBeCalledWith(
       entityId,
       expect.anything(),
