+++ packages/synapse-react-client/test/utils/hooks/SynapseAPI/useFavorites.test.tsx
@@ -1,22 +1,9 @@
-import { renderHook } from '@testing-library/react'
-import React from 'react'
+import { renderHook, waitFor } from '@testing-library/react'
 import { useGetFavorites } from '../../../../src/synapse-queries/user/useFavorites'
 import { EntityHeader, PaginatedResults } from '@sage-bionetworks/synapse-types'
 import { MOCK_CONTEXT_VALUE } from '../../../../mocks/MockSynapseContext'
-import { QueryClient } from 'react-query'
-import { SynapseContextProvider } from '../../../../src/utils/context/SynapseContext'
 import SynapseClient from '../../../../src/synapse-client'
-import FullContextProvider from '../../../../src/utils/context/FullContextProvider'
-const queryClient = new QueryClient()
-
-const wrapper = (props: { children: React.ReactChildren }) => (
-  <FullContextProvider
-    synapseContext={MOCK_CONTEXT_VALUE}
-    queryClient={queryClient}
-  >
-    {props.children}
-  </FullContextProvider>
-)
+import { createWrapper } from '../../../testutils/TestingLibraryUtils'
 
 const expected: PaginatedResults<EntityHeader> = {
   results: [
@@ -41,14 +28,12 @@ const mockGetUserFavorites = jest
   .mockResolvedValue(expected)
 
 describe('useFavorites functionality', () => {
-  beforeEach(() => {
-    queryClient.clear()
-  })
-
   it('correctly calls SynapseClient', async () => {
-    const { result, waitFor } = renderHook(() => useGetFavorites(), { wrapper })
+    const { result } = renderHook(() => useGetFavorites(), {
+      wrapper: createWrapper(),
+    })
 
-    await waitFor(() => result.current.isSuccess)
+    await waitFor(() => expect(result.current.isSuccess).toBe(true))
 
     expect(mockGetUserFavorites).toBeCalledWith(
       MOCK_CONTEXT_VALUE.accessToken,
